F1-BigInteger

+++ src/main/java/com/passingtest/exception/ObjectNotFoundException.java
@@ -1,10 +1,12 @@
 package com.passingtest.exception;
 
+import java.math.BigInteger;
+
 import static java.lang.String.format;
 
 public class ObjectNotFoundException extends RuntimeException {
 
-    public ObjectNotFoundException(int id, Class classParam) {
+    public ObjectNotFoundException(BigInteger id, Class classParam) {
         super(format("Could not find %s %d", classParam.getSimpleName(), id));
     }
 }
