+++ src/main/java/com/passingtest/repository/TestRepository.java
@@ -3,6 +3,8 @@ package com.passingtest.repository;
 import com.passingtest.model.entity.Test;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TestRepository extends CrudRepository<Test, Integer> {
+import java.math.BigInteger;
+
+public interface TestRepository extends CrudRepository<Test, BigInteger> {
 
 }
